move components to ConnectedWalletModal dir
